Update Optimisation using fPortfolio.R
Changed the historical data that was being called from a range of `380 days from the current date to the current date` to `381 days from the current date to the previous day`.
This allows the script to always be available to run. Previously you had to wait for the closing prices of foreign exchange currencies which tends to happen in a window between 22:00 and 23;59 SAST.
In other words, that was the only period in which the script would run smoothly.